ml-service/requirements.txt
# Core ML Libraries - Python 3.13 compatible versions
scikit-learn>=1.4.0
pandas>=2.2.0
numpy>=1.26.0
scipy>=1.12.0

# Deep Learning (optional - for advanced models)
tensorflow==2.15.0
keras==2.15.0

# Time Series Analysis
statsmodels==0.14.1
prophet==1.1.5
tslearn==0.6.3

# Data Processing
openpyxl==3.1.2
xlrd==2.0.1
python-dateutil==2.8.2

# Visualization (optional)
matplotlib>=3.8.2
seaborn>=0.13.0
plotly==5.17.0

# Database & API
supabase>=2.3.4
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2

# Utilities
python-dotenv>=1.0.0
joblib>=1.3.2
tqdm>=4.66.1
requests>=2.31.0

# Model Persistence
pickle-mixin==1.0.2
h5py==3.10.0

# Testing
pytest>=7.4.3
pytest-asyncio==0.21.1
